+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- fix merge groups for events ([#105](https://github.com/PostHog/posthog-android/pull/105))
+
 ## 3.1.11 - 2024-02-28
 
 - fix back compatibility with Kotlin 1.7 ([#104](https://github.com/PostHog/posthog-android/pull/104))